release job fixes
